# docs/openapi.yml
openapi: '3.1.0'
info:
  title: Civic Influence Graph API
  version: 0.1.0
  description: |
    RESTful JSON API for the Civic Influence Graph platform.
    All responses include provenance metadata. No identity-based targeting filters
    are exposed. Authentication required only for write operations and bulk exports.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: /api/v1
    description: Default API base path

# ------------------------------------------------------------------
# Common components
# ------------------------------------------------------------------
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    apiKeyAuth:
      type: http
      scheme: bearer
      description: API key passed as Bearer token
  schemas:
    PaginationMeta:
      type: object
      properties:
        request_id: { type: string, format: uuid }
        timestamp: { type: string, format: date-time }
        data_snapshot: { type: string, nullable: true, format: date-time }
        total_count: { type: integer }
        page: { type: integer }
        page_size: { type: integer }
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            request_id: { type: string, format: uuid }
...
    DashboardData:
      type: object
      properties:
        entity:
          type: object
          properties:
            id: { type: string, format: uuid }
            canonical_name: { type: string }
            entity_type: { type: string }
        funding_summary:
          type: object
          properties:
            total_received: { type: number }
            total_given: { type: number }
            by_sector: { type: array, items: { type: object } }
            top_counterparties: { type: array, items: { type: object } }
        lobbying_summary:
          type: object
          properties:
            engagements_mentioning: { type: integer }
            top_clients: { type: array, items: { type: object } }
            top_issues: { type: array, items: { type: string } }
        voting_summary:
          type: object
          properties:
            total_votes: { type: integer }
            by_party_alignment:
              type: object
              properties:
                with_party: { type: integer }
                against_party: { type: integer }
            recent_votes: { type: array, items: { type: object } }
    Donation:
      type: object
      properties:
        id: { type: string, format: uuid }
        source_entity: { type: object }
        destination_entity: { type: object }
        amount: { type: number }
        transaction_date: { type: string, format: date }
        transaction_type: { type: string, nullable: true }
        election_cycle: { type: string }
        filing_id: { type: string, nullable: true }
        source_system: { type: string }
        source_record_id: { type: string, nullable: true }
...
paths:
...
  /entities/{id}:
    get:
      summary: Get entity by ID
      tags: [Entities]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Entity detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: object }
                  meta: { $ref: '#/components/schemas/PaginationMeta' }
        '404': { description: Not found, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
  /entities/{id}/dashboard:
    get:
      summary: Dashboard aggregation for entity
      tags: [Entities]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: start_date
          in: query
          schema: { type: string, format: date }
        - name: end_date
          in: query
          schema: { type: string, format: date }
      responses:
        '200':
          description: Dashboard data
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/DashboardData' }
                  meta: { $ref: '#/components/schemas/PaginationMeta' }
  /entities/{id}/donations:
    get:
      summary: Paginated donation records
      tags: [Entities]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: direction
          in: query
          schema: { type: string, enum: [received, given] }
        - name: start_date
          in: query
          schema: { type: string, format: date }
        - name: end_date
          in: query
          schema: { type: string, format: date }
        - name: min_amount
          in: query
          schema: { type: number }
        - name: sector
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: page_size
          in: query
          schema: { type: integer, default: 20, maximum: 500 }
      responses:
        '200':
          description: Donation records
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      donations:
                        type: array
                        items: { $ref: '#/components/schemas/Donation' }
                  meta: { $ref: '#/components/schemas/PaginationMeta' }
  /entities/{id}/lobbying:
    get:
      summary: Lobbying engagements for entity
      tags: [Entities]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: page_size
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: Lobbying engagements
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: object }
                  meta: { $ref: '#/components/schemas/PaginationMeta' }
  /entities/{id}/votes:
    get:
      summary: Roll-call votes for legislator
      tags: [Entities]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: page_size
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: Vote records
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: object }
                  meta: { $ref: '#/components/schemas/PaginationMeta' }
...